requirements.txt
numpy==1.21.6
tensorflow==2.8.2
google-cloud-storage
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
matplotlib
Keras-Preprocessing==1.1.2
Pillow==9.1.1
gunicorn==20.1.0

